requirements.txt
asgiref==3.3.1
backcall==0.2.0
beautifulsoup4==4.9.3
colorama==0.4.4
decorator==4.4.2
Django~=3.1.3
django-autoslug==1.9.8
django-bootstrap4==2.3.1
django-crispy-forms==1.11.0
django-model-utils==4.1.0
ipython~=7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.14
psycopg2==2.8.6
Pygments==2.7.4
pytz~=2020.5
soupsieve==2.2
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5

djangorestframework~=3.12.2
cffi~=1.14.4
pip~=21.0.1
cryptography~=3.3.1
py~=1.10.0
six~=1.15.0
packaging~=20.9
setuptools~=49.2.1
pytest~=6.1.2
requests~=2.25.1
django-braces~=1.14.0